Bypass broken Instagram auth
By adding a method in the ApplicationHelper and requiring it, bypass the Instagram
authentication whenever a user registers. Also hide the links that sent to 'Your recent
clips' as this functionality, since Insta auth is broken, were crashing.
Wait for Instagram possible approval of the app to change it back.